PARTIALLY-WORKING! Implement hackish support for regtest network.
I haven't gotten a chance to test it on testnet and mainnet yet, and even if
it does, mainnet vs. regtest net is only differentiated via a dirty hack checking
if there are any connections to other nodes, which probably means we'll be in regtest
mode for a second on bootup prior to our first connections being made. (Not 100% sure
about that though, as I can't currently test it.)

Mostly addresses #45.